requirements.txt
# upgrading this will break some packages
setuptools==65.5.1
confluent_kafka==2.4.0
# dont upgrade flask
#Flask==2.1.2
pytest==7.0.1
python-json-logger==2.0.4
requests==2.31.0
PyHamcrest==2.0.4
flask-restx==1.2.0
bcrypt==3.2.2
PyJWT==2.4.0
PyFunctional==1.4.3
circus==0.17.1
boto3==1.24.56
Flask-Cors==3.0.10
# upgrading Werkzeug breaks request parsing so keep the old version
#Werkzeug==2.1.2
XlsxWriter==3.0.3
mongoengine==0.24.2
jsonschema==4.13.0
python-slugify==6.1.2
mock==4.0.3
celery==5.3.1
#pycurl==7.45.2
text-unidecode ==1.3
fastparquet==0.8.2
Jinja2==3.1.2
beautifulsoup4==4.11.1
pyopenssl==23.1.0
paramiko==2.12.0
openpyxl==3.1.2
newrelic==9.1.0
pandas==2.0.3
pylint==3.0.2
pydantic==1.10
deepdiff==6.6.1
pre-commit==3.5.0
GitPython==3.1.40
unidiff==0.7.5
apache-airflow==2.6.1
apache-airflow-providers-celery==3.3.4
connexion==2.14.2
#pendulum==2.1.2
Flask-Session==0.5.0